+++ crates/storage_models/src/merchant_account.rs
@@ -37,7 +37,7 @@ pub struct MerchantAccount {
     pub api_key: Option<StrongSecret<String>>,
     pub created_at: time::PrimitiveDateTime,
     pub modified_at: time::PrimitiveDateTime,
-    pub intent_fulfillment_time: Option<i32>,
+    pub intent_fulfillment_time: Option<i64>,
 }
 
 #[derive(Clone, Debug, Default, Insertable, router_derive::DebugAsDisplay)]
@@ -59,7 +59,7 @@ pub struct MerchantAccountNew {
     pub routing_algorithm: Option<serde_json::Value>,
     pub primary_business_details: serde_json::Value,
     pub api_key: Option<StrongSecret<String>>,
-    pub intent_fulfillment_time: Option<i32>,
+    pub intent_fulfillment_time: Option<i64>,
 }
 
 #[derive(Debug)]
@@ -79,7 +79,7 @@ pub enum MerchantAccountUpdate {
         metadata: Option<pii::SecretSerdeValue>,
         routing_algorithm: Option<serde_json::Value>,
         primary_business_details: Option<serde_json::Value>,
-        intent_fulfillment_time: Option<i32>,
+        intent_fulfillment_time: Option<i64>,
     },
     StorageSchemeUpdate {
         storage_scheme: storage_enums::MerchantStorageScheme,
@@ -105,7 +105,7 @@ pub struct MerchantAccountUpdateInternal {
     routing_algorithm: Option<serde_json::Value>,
     primary_business_details: Option<serde_json::Value>,
     modified_at: Option<time::PrimitiveDateTime>,
-    intent_fulfillment_time: Option<i32>,
+    intent_fulfillment_time: Option<i64>,
 }
 
 impl From<MerchantAccountUpdate> for MerchantAccountUpdateInternal {
